services/media-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
psycopg2-binary==2.9.9
pydantic==2.8.2
